fix: fixup unit tests
