+++ README.md
@@ -14,16 +14,14 @@ World websocket support is being considered.
 :warning: The API is not yet stable at this time. Use at your own risk. I am
 currently not touching the version number.
 
-| Crate                                  | Usage                                                                                                                                                                                                  |
-| -------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
-| `insim`                                | Connection and protocol implementation.                                                                                                                                                                |
-| `insim_extras`                         | Utilities to make using insim easier.                                                                                                                                                                  |
-| `insim_core`                           | Contains core types shared across other crates.                                                                                                                                                        |
-| `insim_derive`                         | Contains proc_macros for insim_core.                                                                                                                                                                   |
-| `insim_game_data`                      | Contains track information.                                                                                                                                                                            |
-| `insim_pth`                            | Implements a PTH file read/writer.                                                                                                                                                                     |
-| `insim_smx`                            | Implements a SMX file reader/writer.                                                                                                                                                                   |
-| `multi_index` and `multi_index_derive` | `multi_index` started out life as a fork of [lun3x/multi_index_map](https://github.com/lun3x/multi_index_map) with the intention to resubmit upstream. See multi_index/README.md for more information. |
+| Crate             | Usage                                           |
+| ----------------- | ----------------------------------------------- |
+| `insim`           | Connection and protocol implementation.         |
+| `insim_core`      | Contains core types shared across other crates. |
+| `insim_derive`    | Contains proc_macros for insim_core.            |
+| `insim_game_data` | Contains track information.                     |
+| `insim_pth`       | Implements a PTH file read/writer.              |
+| `insim_smx`       | Implements a SMX file reader/writer.            |
 
 ## Documentation
