+++ insim/src/connection/mod.rs
@@ -1,33 +1,25 @@
 //! Connection maintains a connection and provides a Stream and Sink of
 //! [Packets](crate::packets::Packet).
 
-use if_chain::if_chain;
+mod options;
+mod r#type;
 
-pub mod tcp;
-pub mod traits;
-pub mod udp;
-
-pub mod options;
-pub mod transport;
+pub use options::{ConnectionOptions, ReconnectOptions};
 
 use crate::{
-    connection::options::ConnectionOptions,
-    error::{self, Error},
-    packets::{
-        insim::{Isi, Tiny, TinyType, Version},
-        Packet, VERSION,
-    },
+    error::Error,
+    packets::Packet,
     result::Result,
+    tools::{handshake, maybe_keepalive},
+    traits::{ReadPacket, WritePacket},
 };
 
-use insim_core::identifiers::RequestId;
 use std::time::Duration;
+
 use tokio::{
     sync::{broadcast, mpsc, oneshot},
     time,
 };
-
-use self::traits::{ReadPacket, ReadWritePacket, WritePacket};
 
 const TIMEOUT_SECS: u64 = 90;
 
@@ -76,10 +68,7 @@ impl Connection {
     }
 }
 
-async fn run_actor(
-    options: &ConnectionOptions,
-    mut rx: mpsc::Receiver<Command>,
-) -> crate::result::Result<()> {
+async fn run_actor(options: &ConnectionOptions, mut rx: mpsc::Receiver<Command>) -> Result<()> {
     let mut attempt: u64 = 0;
 
     loop {
@@ -147,82 +136,4 @@ async fn run_actor(
             }
         }
     }
-}
-
-pub async fn handshake<I: ReadWritePacket>(
-    inner: &mut I,
-    timeout: Duration,
-    isi: Isi,
-    wait_for_pong: bool,
-    verify_version: bool,
-) -> Result<()> {
-    time::timeout(timeout, inner.write(isi.into())).await??;
-
-    time::timeout(timeout, verify(inner, wait_for_pong, verify_version)).await?
-}
-
-/// Handle the verification of a Transport.
-/// Is Insim server responding the correct version?
-/// Have we received an initial ping response?
-async fn verify<I: ReadWritePacket>(
-    inner: &mut I,
-    verify_version: bool,
-    wait_for_pong: bool,
-) -> Result<()> {
-    if wait_for_pong {
-        // send a ping!
-        inner
-            .write(
-                Tiny {
-                    reqi: RequestId(2),
-                    subt: TinyType::Ping,
-                }
-                .into(),
-            )
-            .await?;
-    }
-
-    let mut received_vers = !verify_version;
-    let mut received_tiny = !wait_for_pong;
-
-    while !received_tiny && !received_vers {
-        match inner.read().await? {
-            None => {
-                return Err(error::Error::Disconnected);
-            }
-            Some(Packet::Tiny(_)) => {
-                received_tiny = true;
-            }
-            Some(Packet::Version(Version { insimver, .. })) => {
-                if insimver != VERSION {
-                    return Err(error::Error::IncompatibleVersion(insimver));
-                }
-
-                received_vers = true;
-            }
-            Some(m) => {
-                /* not the droids we're looking for */
-                tracing::info!("received packet whilst waiting for version and/or ping: {m:?}");
-            }
-        }
-    }
-
-    Ok(())
-}
-
-pub async fn maybe_keepalive<I: ReadWritePacket>(inner: &mut I, packet: &Packet) -> Result<()> {
-    if_chain! {
-        if let Packet::Tiny(i) = &packet;
-        if i.is_keepalive();
-        then {
-            let pong = Tiny{
-                subt: TinyType::None,
-                ..Default::default()
-            };
-
-            inner.write(pong.into()).await?
-        }
-    }
-
-    Ok(())
 }
