+++ insim/src/result.rs
@@ -1,4 +1,2 @@
-use crate::error::Error;
-
 /// A specialized [`Result`] type for insim.
-pub type Result<T> = std::result::Result<T, Error>;
+pub type Result<T> = std::result::Result<T, crate::error::Error>;
