+++ insim/src/tcp.rs
@@ -8,7 +8,7 @@ use crate::error::Error;
 use crate::packets::Packet;
 use crate::result::Result;
 
-use crate::connection::traits::{ReadPacket, ReadWritePacket, WritePacket};
+use crate::traits::{ReadPacket, ReadWritePacket, WritePacket};
 
 pub struct Tcp {
     inner: TcpStream,
