--- multi_index/README.md
@@ -1,16 +1,0 @@
-# multi_index
-
-`multi_index` started out life as a fork of
-[lun3x/multi_index_map](https://github.com/lun3x/multi_index_map) with the
-intention to resubmit upstream.
-
-As time passed by it has become less and less like the original, and is
-fundamentally no longer compatible due to a different api, parts were replaced
-by darling, etc. and it has been split into `multi_index` and
-`multi_index_derive`.
-
-:warning: Warning!
-
-This fork is intended for usage with insim.rs and patina specifically and as
-such it's api, features and implementation should be considered absolutely
-unstable.
