--- multi_index/src/lib.rs
@@ -1,5 +1,0 @@
-#[doc(hidden)]
-pub extern crate slab;
-
-#[doc(inline)]
-pub use multi_index_derive::MultiIndex;
