--- multi_index_derive/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "multi_index_derive"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-syn = { version = "1.0", features = ["full"] }
-quote = { version = "1.0" }
-proc-macro2 = { version = "1.0"}
-proc-macro-error = "1.0.4"
-convert_case = { version = "0.6" }
-darling = "0.14.2"
-
-[lib]
-proc-macro = true
-bench = false
-doctest = false
