--- multi_index_derive/README.md
@@ -1,12 +1,0 @@
-# multi_index_derive
-
-`multi_index_derive` started out life as a fork of
-[lun3x/multi_index_map](https://github.com/lun3x/multi_index_map).
-
-Please see the adjacent `multi_index` directory for more information.
-
-:warning: Warning!
-
-This fork is intended for usage with insim.rs and patina specifically and as
-such it's api, features and implementation should be considered absolutely
-unstable.
